First proper commit of portfolio site
This is the first proper commit of my portfolio website. It has the majority of the content that I want to display and a decent enough layout given my lack of design skills.

Removed unused zip file with an old version of the site that was being used to test the webserver configuration on the host VPS.

Sample pages generated by the Blazor template have been removed.

nginx and application service config files have been moved into their own directory.